Telah menyelesaikan materi bulk delete data

+++ app/Livewire/Employee.php
@@ -18,7 +18,8 @@ class Employee extends Component
         $alamat,
         $updateData = false,
         $employee_id,
-        $katakunci;
+        $katakunci,
+        $employee_selected_id = [];
 
     public function store()
     {
@@ -75,19 +76,30 @@ class Employee extends Component
         $this->alamat = '';
         $this->updateData = false;
         $this->employee_id = '';
+        $this->employee_selected_id = [];
     }
 
     public function delete()
     {
-        $id = $this->employee_id;
-        ModelsEmployee::find($id)->delete();
+        if ($this->employee_id) {
+            $id = $this->employee_id;
+            ModelsEmployee::find($id)->delete();
+        }
+        if (count($this->employee_selected_id)) {
+            for ($x = 0; $x < count($this->employee_selected_id); $x++) {
+                ModelsEmployee::find($this->employee_selected_id[$x])->delete();
+            }
+        }
+
         $this->clear();
         redirect()->with('berhasil', 'Data berhasil di hapus.');
     }
 
     public function delete_confirmation($id)
     {
-        $this->employee_id = $id;
+        if ($id) {
+            $this->employee_id = $id;
+        }
     }
 
     public function render()
